@igorbrasileiro 2023-11-21
I'm facing issues to run lerna, that's hanging. @heath-freenome or @zxbodya can you help me?

![image](https://github.com/rjsf-team/react-jsonschema-form/assets/15680320/76f892b4-ce24-4aaf-89fb-5d0177703e91)

```sh
node: v18.12.1
npm: 9.4.0
```