@heath-freenome 2023-11-22
> I'm facing issues to run lerna, that's hanging. @heath-freenome or @zxbodya can you help me?
> 
> ![image](https://user-images.githubusercontent.com/15680320/284629822-76f892b4-ce24-4aaf-89fb-5d0177703e91.png)
> 
> Versions node: v18.12.1 npm: 9.4.0
> 
> OS: macOS 13.4.1 - m1 pro

Do the other commands (i.e. `npm run build`, `npm run test`, etc) work?