@igorbrasileiro 2023-11-22
> Do the other commands (i.e. `npm run build`, `npm run test`, etc) work?

No, they don't. The script for each package works ok, but from lerna doesn't.