@igorbrasileiro 2023-11-22
Related to: https://github.com/rjsf-team/react-jsonschema-form/issues/1961#issuecomment-1817157331